apps/requirements.txt
Django==5.0
djangorestframework==3.15
pymongo==4.6.2
requests==2.31.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
openai==1.25.0
langchain-community==0.0.36